infrastructure/requirements.txt
aws-cdk-lib>=2.120.0
constructs>=10.3.0 